[inner] knn
