+++ algorithms/kernel/multiclassclassifier/inner/multiclassclassifier_predict_batch_container_v1.h
@@ -1,4 +1,4 @@
-/* file: multiclassclassifier_predict_batch_container.h */
+/* file: multiclassclassifier_predict_batch_container_v1.h */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -38,13 +38,13 @@ namespace multi_class_classifier
 {
 namespace prediction
 {
-namespace interface2
+namespace interface1
 {
 
 template<typename algorithmFPType, prediction::Method pmethod, training::Method tmethod, CpuType cpu>
 BatchContainer<algorithmFPType, pmethod, tmethod, cpu>::BatchContainer(daal::services::Environment::env *daalEnv)
 {
-    __DAAL_INITIALIZE_KERNELS(internal::MultiClassClassifierPredictKernel, pmethod, tmethod, algorithmFPType, classifier::prediction::interface2::Batch, multi_class_classifier::interface2::Parameter);
+    __DAAL_INITIALIZE_KERNELS(internal::MultiClassClassifierPredictKernel, pmethod, tmethod, algorithmFPType, classifier::prediction::interface1::Batch, multi_class_classifier::interface1::Parameter);
 }
 
 template<typename algorithmFPType, prediction::Method pmethod, training::Method tmethod, CpuType cpu>
@@ -57,7 +57,7 @@ template<typename algorithmFPType, prediction::Method pmethod, training::Method 
 services::Status BatchContainer<algorithmFPType, pmethod, tmethod, cpu>::compute()
 {
     const classifier::prediction::Input *input = static_cast<const classifier::prediction::Input *>(_in);
-    classifier::prediction::Result *result = static_cast<classifier::prediction::Result *>(_res);
+    classifier::prediction::interface1::Result *result = static_cast<classifier::prediction::interface1::Result *>(_res);
 
     const NumericTable *a = static_cast<NumericTable *>(input->get(classifier::prediction::data).get());
     const multi_class_classifier::Model *m = static_cast<const multi_class_classifier::Model *>(input->get(classifier::prediction::model).get());
@@ -67,7 +67,7 @@ services::Status BatchContainer<algorithmFPType, pmethod, tmethod, cpu>::compute
     const daal::algorithms::Parameter *par = _par;
     daal::services::Environment::env &env = *_env;
     __DAAL_CALL_KERNEL(env, internal::MultiClassClassifierPredictKernel,
-                       __DAAL_KERNEL_ARGUMENTS(pmethod, tmethod, algorithmFPType, classifier::prediction::interface2::Batch, multi_class_classifier::interface2::Parameter), compute, a, m, r[0], par);
+                       __DAAL_KERNEL_ARGUMENTS(pmethod, tmethod, algorithmFPType, classifier::prediction::interface1::Batch, multi_class_classifier::interface1::Parameter), compute, a, m, r[0], par);
 }
 
 }
