+++ algorithms/kernel/multiclassclassifier/inner/multiclassclassifier_train_batch_container_v1.h
@@ -1,4 +1,4 @@
-/* file: multiclassclassifier_train_batch_container.h */
+/* file: multiclassclassifier_train_batch_container_v1.h */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -37,13 +37,13 @@ namespace multi_class_classifier
 {
 namespace training
 {
-namespace interface2
+namespace interface1
 {
 
 template<typename algorithmFPType, Method method, CpuType cpu>
 BatchContainer<algorithmFPType, method, cpu>::BatchContainer(daal::services::Environment::env *daalEnv)
 {
-    __DAAL_INITIALIZE_KERNELS(internal::MultiClassClassifierTrainKernel, method, algorithmFPType, classifier::training::Batch, multi_class_classifier::Parameter);
+    __DAAL_INITIALIZE_KERNELS(internal::MultiClassClassifierTrainKernel, method, algorithmFPType, classifier::training::interface1::Batch, multi_class_classifier::interface1::Parameter);
 }
 
 template<typename algorithmFPType, Method method, CpuType cpu>
@@ -55,7 +55,7 @@ BatchContainer<algorithmFPType, method, cpu>::~BatchContainer()
 template<typename algorithmFPType, Method method, CpuType cpu>
 services::Status BatchContainer<algorithmFPType, method, cpu>::compute()
 {
-    const classifier::training::Input *input = static_cast<const classifier::training::Input *>(_in);
+    const classifier::training::Input * const input = static_cast<classifier::training::Input *>(_in);
     Result *result = static_cast<Result *>(_res);
 
     const NumericTable *a[2];
@@ -63,9 +63,9 @@ services::Status BatchContainer<algorithmFPType, method, cpu>::compute()
     a[1] = static_cast<NumericTable *>(input->get(classifier::training::labels).get());
     multi_class_classifier::Model *r = static_cast<multi_class_classifier::Model *>(result->get(classifier::training::model).get());
 
-    const multi_class_classifier::Parameter *par = static_cast<const multi_class_classifier::Parameter *>(_par);
+    const multi_class_classifier::interface1::Parameter *par = static_cast<const multi_class_classifier::interface1::Parameter *>(_par);
     daal::services::Environment::env &env = *_env;
-    __DAAL_CALL_KERNEL(env, internal::MultiClassClassifierTrainKernel, __DAAL_KERNEL_ARGUMENTS(method, algorithmFPType, classifier::training::Batch, multi_class_classifier::Parameter), compute, a[0], a[1], r,
+    __DAAL_CALL_KERNEL(env, internal::MultiClassClassifierTrainKernel, __DAAL_KERNEL_ARGUMENTS(method, algorithmFPType, classifier::training::interface1::Batch, multi_class_classifier::interface1::Parameter), compute, a[0], a[1], r,
                        par);
 }
