+++ algorithms/kernel/multiclassclassifier/inner/multiclassclassifier_train_oneagainstone_batch_fpt_cpu_v1.cpp
@@ -1,4 +1,4 @@
-/* file: multiclassclassifier_predict_votebased_batch_fpt_cpu.cpp */
+/* file: multiclassclassifier_train_oneagainstone_batch_fpt_cpu_v1.cpp */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -17,14 +17,15 @@
 
 /*
 //++
-//  Implementation of Wu method for Multi-class classifier
-//  prediction algorithm.
+//  Implementation of One-Against-One method for Multi-class classifier
+//  training algorithm.
 //--
 */
 
-#include "multiclassclassifier_predict_batch_container.h"
-#include "multiclassclassifier_predict_kernel.h"
-#include "multiclassclassifier_predict_votebased_impl.i"
+#include "multiclassclassifier_train_batch_container_v1.h"
+#include "multiclassclassifier_train_kernel.h"
+#include "multiclassclassifier_train_oneagainstone_kernel.h"
+#include "multiclassclassifier_train_oneagainstone_impl.i"
 
 namespace daal
 {
@@ -32,17 +33,23 @@ namespace algorithms
 {
 namespace multi_class_classifier
 {
-namespace prediction
+namespace training
 {
-namespace interface2
+namespace interface1
 {
-template class BatchContainer<DAAL_FPTYPE, voteBased, training::oneAgainstOne, DAAL_CPU>;
+template class BatchContainer<DAAL_FPTYPE, oneAgainstOne,    DAAL_CPU>;
 }
 namespace internal
 {
-template class MultiClassClassifierPredictKernel<voteBased, training::oneAgainstOne, DAAL_FPTYPE, classifier::prediction::Batch, multi_class_classifier::Parameter, DAAL_CPU>;
+
+template class MultiClassClassifierTrainKernel<oneAgainstOne,    DAAL_FPTYPE, classifier::training::interface1::Batch, multi_class_classifier::interface1::Parameter, DAAL_CPU>;
+
 } // namespace internal
-} // namespace prediction
+
+} // namespace training
+
 } // namespace multi_class_classifier
+
 } // namespace algorithms
+
 } // namespace daal
