+++ algorithms/kernel/multiclassclassifier/multiclassclassifier_predict_mccwu_batch_fpt_cpu.cpp
@@ -35,17 +35,12 @@ namespace multi_class_classifier
 {
 namespace prediction
 {
-namespace interface1
-{
-template class BatchContainer<DAAL_FPTYPE, multiClassClassifierWu, training::oneAgainstOne, DAAL_CPU>;
-}
 namespace interface2
 {
 template class BatchContainer<DAAL_FPTYPE, multiClassClassifierWu, training::oneAgainstOne, DAAL_CPU>;
 }
 namespace internal
 {
-template class MultiClassClassifierPredictKernel<multiClassClassifierWu, training::oneAgainstOne, DAAL_FPTYPE, classifier::prediction::interface1::Batch, multi_class_classifier::interface1::Parameter, DAAL_CPU>;
 template class MultiClassClassifierPredictKernel<multiClassClassifierWu, training::oneAgainstOne, DAAL_FPTYPE, classifier::prediction::Batch, multi_class_classifier::Parameter, DAAL_CPU>;
 } // namespace internal
 } // namespace prediction
